+++ src/main/fulcro/easy_server.clj
@@ -13,7 +13,7 @@
     [ring.middleware.gzip :refer [wrap-gzip]]
     [ring.middleware.not-modified :refer [wrap-not-modified]]
     [ring.middleware.resource :refer [wrap-resource]]
-    [ring.util.response :as rsp :refer [response file-response resource-response]] )
+    [ring.util.response :as rsp :refer [response file-response resource-response]])
   (:gen-class))
 
 (defn index [req]
@@ -33,12 +33,12 @@
 
 (defn app-namify-api [default-routes app-name]
   (if-not app-name default-routes
-    (update default-routes 1
-      (fn [m]
-        (let [api-val (get m default-api-key)]
-          (-> m
-            (dissoc default-api-key)
-            (assoc (str "/" app-name default-api-key) api-val)))))))
+                   (update default-routes 1
+                     (fn [m]
+                       (let [api-val (get m default-api-key)]
+                         (-> m
+                           (dissoc default-api-key)
+                           (assoc (str "/" app-name default-api-key) api-val)))))))
 
 (def default-routes
   ["" {"/"             :index
@@ -67,16 +67,16 @@
                          :parser api-parser
                          :env (assoc om-parsing-env :request req)
                          :app-name app-name))
-        (handler req))))
+      (handler req))))
 
 (defn wrap-extra-routes [dflt-handler {:as extra-routes :keys [routes handlers]} om-parsing-env]
   (if-not extra-routes dflt-handler
-    (do (assert (and routes handlers) extra-routes)
-      (fn [req]
-        (let [match (bidi/match-route routes (:uri req) :request-method (:request-method req))]
-          (if-let [bidi-handler (get handlers (:handler match))]
-            (bidi-handler (assoc om-parsing-env :request req) match)
-            (dflt-handler req)))))))
+                       (do (assert (and routes handlers) extra-routes)
+                           (fn [req]
+                             (let [match (bidi/match-route routes (:uri req) :request-method (:request-method req))]
+                               (if-let [bidi-handler (get handlers (:handler match))]
+                                 (bidi-handler (assoc om-parsing-env :request req) match)
+                                 (dflt-handler req)))))))
 
 (defn not-found-handler []
   (fn [req]
@@ -187,20 +187,17 @@
         (throw e))))
   (stop [this]
     (if-not server this
-      (do (server)
-        (timbre/info "web server stopped.")
-        (assoc this :server nil)))))
+                   (do (server)
+                       (timbre/info "web server stopped.")
+                       (assoc this :server nil)))))
 
 (defn make-web-server
   "Builds a web server with an optional argument that
    specifies which component to get `:middleware` from,
-   defaults to `:handler`."
+   defaults to `:handler`. This component requires that your
+   system has a Config component under the key :config."
   [& [handler]]
-  (component/using
-    (component/using
-      (map->WebServer {})
-      [:config])
-    {:handler (or handler :handler)}))
+  (component/using (map->WebServer {}) {:config :config :handler (or handler :handler)}))
 
 (defn make-fulcro-server
   "Make a new fulcro server.