+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Dramatically improved devcards support: state persistent, eliminated console errors. Must port to new
 functions/macros to use. Legacy `fulcro-app` works better than it did, but it is not recommended and will
 cause cards that render the same app to collide.
+- Worked a lot on the server parts of the devguide. Also fixed some doc strings that still referred to load-data
 
 1.0.0-beta5
 -----------